Refactoring broadcast server package name
