+++ broadcastserver/src/main/java/com/neverwinterdp/broadcast/server/broadcast.properties_example
